// web/platform-admin-web/src/api/generated/model/allowedResourceListItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * AI Base App Platform Service API
 * The AI Base App Platform Service API documentation
 * OpenAPI spec version: 1.0
 */
import type { AllowedResourceListItemActionsItem } from './allowedResourceListItemActionsItem';
import type { AllowedResourceListItemScope } from './allowedResourceListItemScope';
import type { AllowedResourceListItemType } from './allowedResourceListItemType';

export interface AllowedResourceListItem {
  /** 리소스 식별자 */
  id: string;
  /** 리소스 코드 */
  code: string;
  /** 리소스 이름 */
  name: string;
  /** 리소스 유형 */
  type: AllowedResourceListItemType;
  /** 리소스 관리 범위 */
  scope: AllowedResourceListItemScope;
  /** 리소스 경로 */
  path?: string;
  /** 아이콘 */
  icon?: string;
  /** 정렬 순서 */
  sortOrder?: number;
  /** 리소스 액션 목록 */
  actions: AllowedResourceListItemActionsItem[];
  /** 부모 리소스 식별자 */
  parent?: string;
  /** 하위 리소스 목록 */
  children: AllowedResourceListItem[];
}