// web/platform-admin-web/src/api/generated/model/allowedResourceListItemType.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * AI Base App Platform Service API
 * The AI Base App Platform Service API documentation
 * OpenAPI spec version: 1.0
 */

/**
 * 리소스 유형
 */
export type AllowedResourceListItemType = typeof AllowedResourceListItemType[keyof typeof AllowedResourceListItemType];


export const AllowedResourceListItemType = {
  MENU: 'MENU',
  COMPONENT: 'COMPONENT',
} as const;